fix imports after moving write_gct
