+++ pycytominer/cyto_utils/write_gct.py
@@ -10,7 +10,7 @@ https://github.com/broadinstitute/cytominer_scripts/blob/master/write_gct.R
 import csv
 import numpy as np
 import pandas as pd
-from pycytominer.cyto_utils.features import infer_cp_features
+from pycytominer.cyto_utils import infer_cp_features
 
 
 def write_gct(
